requirements.txt
pandas
numpy
nltk

sentence_transformers 

pydash

google.ai.generativelanguage
google-auth-oauthlib

python-dotenv

chromadb

openpyxl

loguru

streamlit
grobid-client-python
beautifulsoup4
pymupdf
lxml